requirements.txt
# ----------------------
# Core
# ----------------------
pandas==2.2.2
numpy==1.26.4

# ----------------------
# Machine Learning
# ----------------------
scikit-learn==1.5.1
scipy==1.13.1   # ✅ required dependency

# ----------------------
# Vector Search
# ----------------------
faiss-cpu==1.7.4   # ✅ more stable for Windows

# ----------------------
# Backend
# ----------------------
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ----------------------
# Frontend
# ----------------------
streamlit==1.37.0

# ----------------------
# Utilities
# ----------------------
requests==2.32.3
python-dotenv==1.0.1
gdown==5.2.0